Move auth classes to a file of their own
