+++ posthog/middleware.py
@@ -6,7 +6,7 @@ from django.core.exceptions import MiddlewareNotUsed
 from django.http import HttpRequest, HttpResponse
 from django.middleware.csrf import CsrfViewMiddleware
 
-from .utils import PersonalAPIKeyAuthentication
+from .auth import PersonalAPIKeyAuthentication
 
 
 class AllowIP(object):
@@ -102,7 +102,7 @@ class CsrfOrKeyViewMiddleware(CsrfViewMiddleware):
     def process_view(self, request, callback, callback_args, callback_kwargs):
         result = super().process_view(request, callback, callback_args, callback_kwargs)  # None if request accepted
         # if super().process_view did not find a valid CSRF token, try looking for a personal API key
-        if result is not None and PersonalAPIKeyAuthentication().find_key_with_source(request) is not None:
+        if result is not None and PersonalAPIKeyAuthentication.find_key_with_source(request) is not None:
             return self._accept(request)
         return result
